move gql constants to a common
